fix: session synchronization and duplicate process_job id

+++ src/main/java/com/ai/learning/backend/controller/UserController.java
@@ -45,26 +45,26 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public UserResponse updateProfile(@PathVariable Integer id,@RequestBody UpdateUserRequest request) {
+    public UserResponse updateProfile(@PathVariable Long id,@RequestBody UpdateUserRequest request) {
         return ApiResponse.<UserResponse>builder().
                 result(userService.updateProfileRequest(id,request)).build().getResult();
     }
 
     @GetMapping("/{id}/can-upload")
-    public boolean canUpload(@PathVariable Integer id) {
+    public boolean canUpload(@PathVariable Long id) {
         return ApiResponse.<Boolean>builder().
                 result(userService.canUpload(id)).build().getResult();
     }
 
     @GetMapping("/{id}/update-usage")
-    public ApiResponse.ApiResponseBuilder<Void> updateUsage(@PathVariable Integer id) {
+    public ApiResponse.ApiResponseBuilder<Void> updateUsage(@PathVariable Long id) {
         userService.updateUsage(id);
         return ApiResponse.<Void>builder().
                 message("Update usage successfully");
     }
 
     @DeleteMapping("/{userId}")
-    public ApiResponse<Void> deleteUser(@PathVariable  Integer userId) {
+    public ApiResponse<Void> deleteUser(@PathVariable Long userId) {
         userService.deleteUser(userId);
         return ApiResponse.<Void>builder()
                 .message("User has been deleted successfully")
