+++ src/main/java/com/ai/learning/backend/repository/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     // Find User by Email
     Optional<User> findByEmail(String email);
 
@@ -22,5 +22,5 @@ public interface UserRepository extends JpaRepository<User,Integer> {
     Optional<User> findByUsername(String username);
 
     //Delete User
-    void deleteByUserId(Integer userId);
+    void deleteByUserId(Long userId);
 }
